Moved share event button into the component file for the event route.
