+++ app/event/eventCard.tsx
@@ -2,7 +2,7 @@ import React from "react";
 
 import DateTimeWithDST from "./dateHandling";
 import AddEventToCalendarButton from "./components/AddEventToCalendarButton";
-import ShareButton from "./shareEventButton";
+import ShareButton from "./components/shareEventButton";
 
 import { Event } from "./Event";
 import OpenLinkInNewWindowButton from "../components/buttons/openLinkInNewWindow";
